+++ src/db/migrations/20210831224850-create-titles-groups.js
@@ -17,7 +17,7 @@ module.exports = {
         }
       },
       season: {
-        type: Sequelize.INTEGER(2),
+        type: Sequelize.INTEGER,
         allowNull: true,
       },
       group_id: {
